Consolidate systematics headers
